requirements.txt
Django
djangorestframework
Pillow
phonenumbers
django-phonenumber-field
